Add necesary fields in program.cs and modify some codes, changed database connection string

+++ FoodRecipeAPI/Dto/RateDto.cs
@@ -6,7 +6,7 @@ namespace FoodRecipeAPI.Dto
     public class RateDto
     {
         public int Id { get; set; }
-        public decimal score { get; set; }
+        public double score { get; set; }
         public string comment { get; set; }
         public int RecipeId { get; set; }
     }
